fix: jitter & g722 issue

+++ src/media/tests/recorder.rs
@@ -55,14 +55,14 @@ async fn test_recorder() -> Result<()> {
         let left_frame = AudioFrame {
             track_id: left_channel_id.clone(),
             samples: Samples::PCM(left_samples),
-            timestamp: (i * 100) as u32, // Increment timestamp by 100ms
+            timestamp: (i * 100), // Increment timestamp by 100ms
             sample_rate: 16000,
         };
 
         let right_frame = AudioFrame {
             track_id: right_channel_id.clone(),
             samples: Samples::PCM(right_samples),
-            timestamp: (i * 100) as u32, // Same timestamp for synchronized channels
+            timestamp: (i * 100), // Same timestamp for synchronized channels
             sample_rate: 16000,
         };
